requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
DBUtils==1.3
Flask==1.0
Flask-Admin==1.5.5
Flask-BabelEx==0.9.4
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mongoengine==0.19.1
multidict==4.7.5
Pillow==7.0.0
pymongo==3.10.1
PyMySQL==0.9.3
pytz==2019.3
qiniu==7.2.2
redis==3.2.0
requests==2.21.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.1
tornado==6.0.4
typing-extensions==3.7.4.1
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
yarl==1.4.2
